refactor topratedmovies to use hooks
